Add files via upload
The unsteady_natural_convection_in_cavity.m script defines the problem parameters, creates the grid, and performs the simulation. The script uses a finite difference method to solve the equations numerically. The simulation is performed until a specified end time, with results written to a video file at regular intervals.

The simulation parameters include the number of grid points, the time step, the end time, the thermal diffusivity, and the kinematic viscosity. The temperature boundary conditions are set such that the left side of the cavity is at a higher temperature than the right side.

The script outputs a video file of the simulation results, as well as the history of the velocity and temperature fields during the simulation.



To run the simulation, simply run the unsteady_natural_convection_in_cavity.m script in MATLAB. The simulation results will be written to the output.mp4 file in the same directory.

References

Ghia, U., Ghia, K.N., and Shin, C.T. (1982). High-Re solutions for incompressible flow using the Navier-Stokes equations and a multigrid method. Journal of Computational Physics, 48(3), 387-411. https://doi.org/10.1016/0021-9991(82)90058-4